model-sequence-router: derive cursor identity from conversation content
The plugin keyed its sequence cursor on the credential-affinity identity exported by core.
That accessor returns only the primary value of a paired extraction, so a conversation that
first carried conversation.id and later added prompt_cache_key changed key from conv: to pck:
and restarted its sequence mid-conversation. The same ladder exposed transport and credential
selection concerns to the routing layer. Cursor identity now comes from the shared protocol
aware conversation derivation, which reads caller scope, leading instructions, and the first
user input across every inbound format. Diagnostic lane snapshots separately accumulated until
reconfiguration, because the periodic loop swept cursors alone.

- Derive conversation identity through sdk/cliproxy/session, so no transport header, cache lane,
  affinity group, or account identifier reaches a cursor key
- Collapse conversationIdentity to a string and compute its diagnostic source at use, because the
  origin is no longer independent state
- Add an expiringStore contract so one periodic sweep releases both cursors and diagnostic lane
  observations under a single interval
- Group the lane store methods beside their type and add size readouts that controlled-clock tests
  observe directly
- Split router and state tests by purpose into cursor probe, observation, and identity files to
  keep each file within the module size limit
- Share one conversation transcript fixture across tests so identity cases exercise the derivation
  rather than request headers
- Raise the plugin version to 0.10.0 for the changed identity contract
- Document YAML-first provider selection, credential-only affinity, and the content-derived cursor
  limitation
- Tidy the plugin module graph after the credential-affinity import was dropped
