+++ examples/plugin/model-sequence-router/go/router_test.go
@@ -1,15 +1,13 @@
 package main
 
 import (
-	"encoding/json"
-	"net/http"
-	"strings"
 	"testing"
 	"time"
 
 	"github.com/router-for-me/CLIProxyAPI/v7/sdk/pluginapi"
 )
 
+// capturedPluginLog records one log entry the plugin emitted through its host.
 type capturedPluginLog struct {
 	level          string
 	message        string
@@ -49,18 +47,20 @@ func TestRouterMatchingSuffixAndSequence(t *testing.T) {
 	if unmatched.Handled {
 		t.Fatalf("unmatched response = %#v", unmatched)
 	}
-	base := pluginapi.ModelRouteRequest{
-		RequestedModel:     "ITERATIVE-MODEL(max)",
-		AvailableProviders: []string{"codex", "claude"},
-		Headers:            http.Header{"Session-Id": {"session-a"}},
+	sequenceTurn := func(turns int) pluginapi.ModelRouteRequest {
+		return pluginapi.ModelRouteRequest{
+			RequestedModel:     "ITERATIVE-MODEL(max)",
+			AvailableProviders: []string{"codex", "claude"},
+			Body:               messagesTurn(t, "sequence", turns),
+		}
 	}
-	first := runtime.route(base)
+	first := runtime.route(sequenceTurn(1))
 	if !first.Handled || first.Target != "codex" || first.TargetModel != "terra(max)" || first.Reason != routeReason {
 		t.Fatalf("first response = %#v", first)
 	}
-	second := runtime.route(base)
-	third := runtime.route(base)
-	fourth := runtime.route(base)
+	second := runtime.route(sequenceTurn(2))
+	third := runtime.route(sequenceTurn(3))
+	fourth := runtime.route(sequenceTurn(4))
 	if second.TargetModel != "terra(max)" || third.TargetModel != "terra(max)" {
 		t.Fatalf("second/third = %#v / %#v", second, third)
 	}
@@ -109,9 +109,10 @@ func TestRouterUsesConfiguredRandomStart(t *testing.T) {
 	req := pluginapi.ModelRouteRequest{
 		RequestedModel:     "iterative-model",
 		AvailableProviders: []string{"codex", "claude"},
-		Headers:            http.Header{"X-Session-ID": {"random-session"}},
+		Body:               messagesTurn(t, "random", 1),
 	}
 	first := runtime.route(req)
+	req.Body = messagesTurn(t, "random", 2)
 	second := runtime.route(req)
 	if first.Target != "claude" || first.TargetModel != "opus(high)" {
 		t.Fatalf("random first route = %#v", first)
@@ -129,22 +130,22 @@ func TestRouterSessionsAndProviderSkipping(t *testing.T) {
 	reqA := pluginapi.ModelRouteRequest{
 		RequestedModel:     "iterative-model",
 		AvailableProviders: []string{"codex", "claude"},
-		Body:               []byte(`{"metadata":{"user_id":"user_x_account__session_aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee"}}`),
+		Body:               messagesTurn(t, "conversation-a", 1),
 	}
 	reqB := reqA
-	reqB.Body = []byte(`{"metadata":{"user_id":"user_x_account__session_ffffffff-bbbb-cccc-dddd-eeeeeeeeeeee"}}`)
+	reqB.Body = messagesTurn(t, "conversation-b", 1)
 	for range 2 {
 		if got := runtime.route(reqA); got.Target != "codex" {
-			t.Fatalf("session A route = %#v", got)
+			t.Fatalf("conversation A route = %#v", got)
 		}
 	}
 	if got := runtime.route(reqB); got.TargetModel != "terra" {
-		t.Fatalf("session B did not start independently: %#v", got)
+		t.Fatalf("conversation B did not start independently: %#v", got)
 	}
 	skip := pluginapi.ModelRouteRequest{
 		RequestedModel:     "iterative-model",
 		AvailableProviders: []string{"claude"},
-		Body:               []byte(`{"conversation_id":"skip-conversation"}`),
+		Body:               messagesTurn(t, "skip-conversation", 1),
 	}
 	if got := runtime.route(skip); got.Target != "claude" {
 		t.Fatalf("provider skip route = %#v", got)
@@ -152,17 +153,20 @@ func TestRouterSessionsAndProviderSkipping(t *testing.T) {
 	if got := runtime.route(pluginapi.ModelRouteRequest{
 		RequestedModel:     "iterative-model",
 		AvailableProviders: []string{"codex", "claude"},
-		Body:               []byte(`{"conversation_id":"skip-conversation"}`),
+		Body:               messagesTurn(t, "skip-conversation", 2),
 	}); got.Target != "codex" {
 		t.Fatalf("skip cursor did not wrap after selected target: %#v", got)
 	}
 }
 
-func TestRouterRecognizesPromptCacheAndMessageHash(t *testing.T) {
+// TestRouterHoldsPositionForUnchangedRequests proves a repeated conversation state
+// keeps its position, while a request carrying no conversation content routes without
+// a cursor.
+func TestRouterHoldsPositionForUnchangedRequests(t *testing.T) {
 	runtime := newTestRuntime(t)
 	for _, body := range [][]byte{
 		[]byte(`{"prompt_cache_key":"cache-a"}`),
-		[]byte(`{"messages":[{"role":"user","content":"stable opening"}]}`),
+		messagesTurn(t, "stable", 1),
 	} {
 		req := pluginapi.ModelRouteRequest{
 			RequestedModel:     "iterative-model",
@@ -188,162 +192,6 @@ func TestRouterStatelessAndUnavailable(t *testing.T) {
 	req.AvailableProviders = nil
 	if got := runtime.route(req); got.Handled {
 		t.Fatalf("unavailable response = %#v", got)
-	}
-}
-
-func TestRouterLogsSelectionThroughHostWithoutSensitiveIdentifiers(t *testing.T) {
-	runtime := newTestRuntime(t)
-	logs := captureRouteLogs(runtime)
-	const sessionID = "private-session-identifier"
-	resp := runtime.routeWithCallback(pluginapi.ModelRouteRequest{
-		RequestedModel:     "iterative-model(max)",
-		AvailableProviders: []string{"codex", "claude"},
-		Headers: http.Header{
-			"X-Session-ID":  {sessionID},
-			"Authorization": {"Bearer private-token"},
-		},
-		Body: []byte(`{"messages":[{"role":"user","content":"private prompt"}]}`),
-	}, "callback-123")
-	if !resp.Handled || len(*logs) != 1 {
-		t.Fatalf("response/logs = %#v / %#v", resp, *logs)
-	}
-	got := (*logs)[0]
-	wantMessage := "model-sequence-router: selected target alias=Iterative-Model position=0 requested=max"
-	if got.level != "debug" || got.message != wantMessage || got.hostCallbackID != "callback-123" {
-		t.Fatalf("log identity = %#v", got)
-	}
-	if got.fields["alias"] != "Iterative-Model" || got.fields["provider"] != "codex" ||
-		got.fields["model"] != "terra(max)" ||
-		got.fields["advanced"] != true || got.fields["random_start"] != true || got.fields["sequence_index"] != 0 ||
-		got.fields["requested_effort"] != "max" {
-		t.Fatalf("log fields = %#v", got.fields)
-	}
-	if got.fields["session_hash"] != shortSessionHash("header:"+sessionID) {
-		t.Fatalf("session hash = %#v", got.fields["session_hash"])
-	}
-	raw, errMarshal := json.Marshal(map[string]any{
-		"level": got.level, "message": got.message, "fields": got.fields, "host_callback_id": got.hostCallbackID,
-	})
-	if errMarshal != nil {
-		t.Fatal(errMarshal)
-	}
-	for _, secret := range []string{sessionID, "private-token", "private prompt"} {
-		if strings.Contains(string(raw), secret) {
-			t.Fatalf("log contains sensitive value %q: %s", secret, raw)
-		}
-	}
-}
-
-// captureRouteLogs redirects plugin logging into a slice the caller inspects.
-func captureRouteLogs(runtime *runtimeState) *[]capturedPluginLog {
-	logs := make([]capturedPluginLog, 0, 4)
-	runtime.logger = func(level, message string, fields map[string]any, hostCallbackID string) {
-		logs = append(logs, capturedPluginLog{
-			level: level, message: message, fields: fields, hostCallbackID: hostCallbackID,
-		})
-	}
-	return &logs
-}
-
-func TestRouterAdvancesOncePerConversationState(t *testing.T) {
-	runtime := newTestRuntime(t)
-	logs := captureRouteLogs(runtime)
-	countRequest := pluginapi.ModelRouteRequest{
-		RequestedModel:     "iterative-model",
-		AvailableProviders: []string{"codex", "claude"},
-		Body:               []byte(`{"messages":[{"role":"user","content":"first turn"}],"max_tokens":1}`),
-	}
-	if got := runtime.route(countRequest); got.Target != "codex" {
-		t.Fatalf("count-shaped call = %#v", got)
-	}
-	generationRequest := countRequest
-	generationRequest.Body = []byte(`{"messages":[{"role":"user","content":"first turn"}],"max_tokens":512}`)
-	generationRequest.Stream = true
-	if got := runtime.route(generationRequest); got.Target != "codex" {
-		t.Fatalf("generation call = %#v", got)
-	}
-	changed := generationRequest
-	changed.Body = []byte(`{"messages":[{"role":"user","content":"first turn"},{"role":"assistant","content":"reply"},{"role":"user","content":"second turn"}]}`)
-	if got := runtime.route(changed); got.Target != "codex" {
-		t.Fatalf("changed state call = %#v", got)
-	}
-	wantOutcomes := []string{"advanced", "replayed", "advanced"}
-	wantIndexes := []int{0, 0, 1}
-	if len(*logs) != len(wantOutcomes) {
-		t.Fatalf("route records = %d, want %d", len(*logs), len(wantOutcomes))
-	}
-	for index, want := range wantOutcomes {
-		record := (*logs)[index]
-		if record.fields["outcome"] != want || record.fields["sequence_index"] != wantIndexes[index] {
-			t.Fatalf("record %d = %#v, want %s at %d", index, record.fields, want, wantIndexes[index])
-		}
-		if record.fields["identity_source"] != string(identitySourceDerived) {
-			t.Fatalf("record %d identity source = %#v", index, record.fields["identity_source"])
-		}
-	}
-}
-
-func TestRouterEmitsOneNoticePerSkippedPosition(t *testing.T) {
-	runtime := newTestRuntime(t)
-	logs := captureRouteLogs(runtime)
-	got := runtime.route(pluginapi.ModelRouteRequest{
-		RequestedModel:     "iterative-model",
-		AvailableProviders: []string{"claude"},
-		Body:               []byte(`{"messages":[{"role":"user","content":"skip"}]}`),
-	})
-	if got.Target != "claude" {
-		t.Fatalf("skip route = %#v", got)
-	}
-	notices := make([]capturedPluginLog, 0, 3)
-	for _, record := range *logs {
-		if record.fields["event"] == "skip" {
-			notices = append(notices, record)
-		}
-	}
-	if len(notices) != 3 {
-		t.Fatalf("skip notices = %d, want 3", len(notices))
-	}
-	for index, notice := range notices {
-		if notice.level != "warn" || notice.fields["provider"] != "codex" || notice.fields["sequence_index"] != index {
-			t.Fatalf("notice %d = %#v", index, notice.fields)
-		}
-	}
-}
-
-func TestErrorPolicyKeepsPositionAndSelfTargets(t *testing.T) {
-	cfg, errCompile := decodeAndCompileConfig([]byte(`
-unavailable_provider: error
-aliases:
-  - alias: held
-    random_start: false
-    targets: [{provider: codex, model: one}, {provider: claude, model: two}]
-`), 1)
-	if errCompile != nil {
-		t.Fatal(errCompile)
-	}
-	runtime := newRuntimeState(func() time.Time { return time.Unix(100, 0) })
-	runtime.config.Store(cfg)
-	logs := captureRouteLogs(runtime)
-	req := pluginapi.ModelRouteRequest{
-		RequestedModel:     "held",
-		AvailableProviders: []string{"claude"},
-		Body:               []byte(`{"messages":[{"role":"user","content":"held turn"}]}`),
-	}
-	held := runtime.route(req)
-	if !held.Handled || held.TargetKind != pluginapi.ModelRouteTargetSelf || held.Reason != unavailableCode {
-		t.Fatalf("held response = %#v", held)
-	}
-	if len(*logs) != 1 {
-		t.Fatalf("held records = %#v", *logs)
-	}
-	record := (*logs)[0]
-	if record.level != "warn" || record.fields["provider"] != "codex" || record.fields["sequence_index"] != 0 {
-		t.Fatalf("held record = %#v", record.fields)
-	}
-	req.AvailableProviders = []string{"codex", "claude"}
-	recovered := runtime.route(req)
-	if recovered.Target != "codex" || recovered.TargetModel != "one" {
-		t.Fatalf("recovered response = %#v, want the kept position", recovered)
 	}
 }
