+++ examples/plugin/model-sequence-router/go/session_identity.go
@@ -1,40 +1,35 @@
 package main
 
 import (
-	"strings"
-
-	coreauth "github.com/router-for-me/CLIProxyAPI/v7/sdk/cliproxy/auth"
+	coreexecutor "github.com/router-for-me/CLIProxyAPI/v7/sdk/cliproxy/executor"
+	coresession "github.com/router-for-me/CLIProxyAPI/v7/sdk/cliproxy/session"
 	"github.com/router-for-me/CLIProxyAPI/v7/sdk/pluginapi"
+	sdktranslator "github.com/router-for-me/CLIProxyAPI/v7/sdk/translator"
 )
 
-const (
-	// coreMessageHashPrefix marks the content hash core returns when a client
-	// supplies no durable session value. That hash folds in the first assistant
-	// reply, so it names one conversation differently before and after its first
-	// answer.
-	coreMessageHashPrefix = "msg:"
+// derivedIdentityPrefix marks an identity computed from conversation content.
+const derivedIdentityPrefix = "derived:"
 
-	// derivedIdentityPrefix marks an identity this plugin computed, keeping it
-	// distinguishable from any value core supplied.
-	derivedIdentityPrefix = "conv:"
-)
-
-// identitySource names where a conversation identity came from.
+// identitySource names how routing treated one request's identity.
 type identitySource string
 
 const (
-	// identitySourceCore marks an identity the client or core supplied durably.
-	identitySourceCore identitySource = "core"
 	// identitySourceDerived marks an identity computed from conversation content.
 	identitySourceDerived identitySource = "derived"
-	// identitySourceAbsent marks a request that carries no identity at all.
+	// identitySourceAbsent marks a request that carries no conversation content.
 	identitySourceAbsent identitySource = "absent"
 )
 
-// conversationIdentity names one conversation and records the origin of its value.
-type conversationIdentity struct {
-	Value  string
-	Source identitySource
+// conversationIdentity names one conversation. An empty identity marks a request
+// carrying no conversation content, which keys no cursor.
+type conversationIdentity string
+
+// source reports how routing treated this identity.
+func (c conversationIdentity) source() identitySource {
+	if c == "" {
+		return identitySourceAbsent
+	}
+	return identitySourceDerived
 }
 
 // turnIdentity names one conversation state. A value exists only when the
@@ -44,38 +39,19 @@ type turnIdentity struct {
 }
 
 // newConversationIdentity resolves the identity that keys one conversation cursor.
-// A durable core identity passes through unchanged. A core content-hash fallback is
-// replaced by an identity built from values that hold constant across every turn.
-func newConversationIdentity(req pluginapi.ModelRouteRequest, observation requestObservation, salt []byte) conversationIdentity {
-	var identity conversationIdentity
-	core := strings.TrimSpace(coreauth.ExtractSessionID(req.Headers, req.Body, req.Metadata))
-	if core != "" && !strings.HasPrefix(core, coreMessageHashPrefix) {
-		identity = conversationIdentity{Value: core, Source: identitySourceCore}
-	} else if derived := derivedConversationIdentity(observation, salt); derived != "" {
-		identity = conversationIdentity{Value: derived, Source: identitySourceDerived}
-	} else {
-		// Neither a durable client value nor conversation content names this request.
-		identity = conversationIdentity{Source: identitySourceAbsent}
+// Sequence position belongs to the conversation, so identity folds the leading
+// instructions and the first complete user input through the shared protocol-aware
+// derivation. Every inbound format resolves one conversation the same way, and
+// transport identifiers naming a cache lane, an affinity group, one request, or an
+// account never reach a cursor key. The identity is empty when the body carries no
+// user input.
+func newConversationIdentity(req pluginapi.ModelRouteRequest) conversationIdentity {
+	callerScope, _ := req.Metadata[coreexecutor.CallerScopeMetadataKey].(string)
+	derived := coresession.DeriveID(sdktranslator.FromString(req.SourceFormat), req.Body, callerScope)
+	if derived == "" {
+		return ""
 	}
-	return identity
-}
-
-// derivedConversationIdentity folds the system fingerprint and the first history
-// item fingerprint into one identity. Both hold constant for the life of a
-// conversation. The result is empty when the request carries neither value.
-func derivedConversationIdentity(observation requestObservation, salt []byte) string {
-	parts := make([]string, 0, 2)
-	if observation.SystemFingerprint != "" {
-		parts = append(parts, observation.SystemFingerprint)
-	}
-	if len(observation.HistoryItems) > 0 {
-		parts = append(parts, observation.HistoryItems[0])
-	}
-	identity := ""
-	if len(parts) > 0 {
-		identity = derivedIdentityPrefix + fingerprintStrings(parts, salt)
-	}
-	return identity
+	return conversationIdentity(derivedIdentityPrefix + derived)
 }
 
 // newTurnIdentity names the conversation state carried by one request. It returns
